+++ src/app/core/services/storage-class-abstraction.abstract.ts
@@ -1,0 +1,3 @@
+export abstract class StorageClassAbstraction {
+  localstorage: any = window.localStorage;
+}
