+++ src/app/pages/compare-resume/compare-resume.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CompareResumeService } from './compare-resume.service';
+
+describe('CompareResumeService', () => {
+  let service: CompareResumeService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(CompareResumeService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
